Modified noise to use doubles instead of floats
This allows for far greater precision if doing procedural generation as noise will now be precise until at least +- int.MaxValue

+++ Mmc.MonoGame.Curves/src/2D/NoiseCurve2D.cs
@@ -44,7 +44,7 @@ namespace Mmc.MonoGame.Curves._2D.Noise
 
             Vector2 originalPoint = ParentCurve.GetPoint(t);
 
-            return originalPoint + Scale * Noise.GetValue(originalPoint.X, originalPoint.Y) * Vector2.Normalize(ParentCurve.GetNormal(t));
+            return originalPoint + Scale * (float)Noise.GetValue(originalPoint.X, originalPoint.Y) * Vector2.Normalize(ParentCurve.GetNormal(t));
         }
 
         public override Vector2 GetTangent(float t)
