+++ Mmc.MonoGame.Noise/src/Combiners/DoubleNoiseCombiner.cs
@@ -6,16 +6,16 @@
 
         private INoise NoiseB { get; init; }
 
-        private Func<INoise, INoise, float> NoiseFunc { get; init; }
+        private Func<INoise, INoise, double> NoiseFunc { get; init; }
 
-        public DoubleNoiseCombiner(INoise noiseA, INoise noiseB, Func<INoise, INoise, float> noiseFunc)
+        public DoubleNoiseCombiner(INoise noiseA, INoise noiseB, Func<INoise, INoise, double> noiseFunc)
         {
             NoiseA = noiseA;
             NoiseB = noiseB;
             NoiseFunc = noiseFunc;
         }
 
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
             return NoiseFunc?.Invoke(NoiseA, NoiseB) ?? NoiseA.GetValue(x, y);
         }
