+++ Mmc.MonoGame.Noise/src/Combiners/GenericNoiseCombiner.cs
@@ -4,15 +4,15 @@
     {
         private INoise[] SourceNoises { get; init; }
 
-        private Func<INoise[], float> NoiseFunc { get; init; }
+        private Func<INoise[], double> NoiseFunc { get; init; }
 
-        public GenericNoiseCombiner(Func<INoise[], float> noiseFunc, params INoise[] sourceNoises)
+        public GenericNoiseCombiner(Func<INoise[], double> noiseFunc, params INoise[] sourceNoises)
         {
             SourceNoises = sourceNoises;
             NoiseFunc = noiseFunc;
         }
 
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
             return NoiseFunc?.Invoke(SourceNoises) ?? (SourceNoises.Length > 0 ? SourceNoises[0].GetValue(x, y) : 0);
         }
