+++ Mmc.MonoGame.Noise/src/Combiners/SelectNoise.cs
@@ -8,9 +8,9 @@
 
         private INoise ControlNoise { get; init; }
 
-        private float Threshold { get; init; }
+        private double Threshold { get; init; }
 
-        public SelectNoise(INoise noiseA, INoise noiseB, INoise controlNoise, float threshold)
+        public SelectNoise(INoise noiseA, INoise noiseB, INoise controlNoise, double threshold)
         {
             NoiseA = noiseA;
             NoiseB = noiseB;
@@ -18,7 +18,7 @@
             Threshold = threshold;
         }
 
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
             return ControlNoise.GetValue(x, y) < Threshold ? NoiseA.GetValue(x, y) : NoiseB.GetValue(x, y);
         }
